Fixed slow program issue. Expand for details.
The new angelo picture was causing the program to be slow. Originally, the back.png picture was around 40kB.
The updated png file with 'with love, angelo' text doubled in size. This took it's toll on the program, as
+40kB adds double the overhead for rendering angelo images (many, not just one).

To solve this, I started from the originally download .jpg file, which was about 20kB.  Then I added text, but not
high quality.  Exported into png to 40kB.  Render many 40kB images is way faster than rendering many 80kB images.
Thus, problem solved and lesson learned:  be careful with image sizes.